requirements.txt
altair
pandas
streamlit
web3==6.0.0
eth_abi
numpy
plotly
requests
logging
binance-connector